+++ Disney-API/Models/Schemes/MovieCreate.cs
@@ -7,6 +7,6 @@ namespace Disney_API.Models.Schemes
         public string Imagen { get; set; } = null!;
         public string Titulo { get; set; } = null!;
         public DateTime Fecha { get; set; }
-        public double Calificacion { get; set; }
+        public float Calificacion { get; set; }
     }
 }
